+++ packages/authorization/user-message.go
@@ -2,18 +2,13 @@
    Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
    License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
 
-package api
+package authorization
 
 import (
-	"github.com/dmalix/financelime-rest-api/packages/authorization"
+	"github.com/dmalix/financelime-rest-api/models"
+	"net/mail"
 )
 
-type Handler struct {
-	service authorization.UserService
-}
-
-func NewHandler(service authorization.UserService) *Handler {
-	return &Handler{
-		service: service,
-	}
+type UserMessage interface {
+	AddEmailMessageToQueue(messageQueue chan models.EmailMessage, to mail.Address, subject, body string, messageID ...string) error
 }